requirements.txt
# PDFBooklet/requirements.txt
PyQt6>=6.5.0
pypdf>=3.0.0
Pillow>=10.0.0
PyMuPDF>=1.23.0